requirements.txt
fastapi==0.115.3
uvicorn==0.32.0
langchain-core==0.3.12
langchain==0.3.4
langchain-community
langchain-openai==0.2.3
twilio==9.3.4
openai==1.52.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
pypdf==5.0.1
python-multipart==0.0.12
onnxruntime==1.19.2
chromadb==0.5.15
rank-bm25==0.2.2
Scrapy==2.11.2
beautifulsoup4==4.12.3
